deployment/bicep/requirements.txt
azure-identity>=1.12.0
azure-keyvault-secrets>=4.7.0
azure-ai-textanalytics>=5.3.0
